Fixing type casts.

+++ server/src/main/java/com/example/server/POJO/OrderRequest.java
@@ -10,9 +10,9 @@ public class OrderRequest {
 
     private List<Integer> amounts;
 
-    private Long shopId;
+    private Integer shopId;
 
-    private Long storageId;
+    private Integer storageId;
 
 
     public String getEmail() {
@@ -39,19 +39,19 @@ public class OrderRequest {
         this.amounts = amounts;
     }
 
-    public Long getShopId() {
+    public Integer getShopId() {
         return shopId;
     }
 
-    public void setShopId(Long shopId) {
+    public void setShopId(Integer shopId) {
         this.shopId = shopId;
     }
 
-    public Long getStorageId() {
+    public Integer getStorageId() {
         return storageId;
     }
 
-    public void setStorageId(Long storageId) {
+    public void setStorageId(Integer storageId) {
         this.storageId = storageId;
     }
 }
